Fixed a minor bug in Huffman compression algorithm
Compression is performed in place.
In case there was no compression at all, the compressed buffer would be
one element larger than the original one.
In that case, a buffer overflow would first happen, then the buffer
would be expanded to the correct size (of one additional element),
filling the additional element with a 0.
The buffer overflow can create any type of problems. If none happens,
then the remaining problem is a 0x0000 as last element instead of a
0x8000, which apparently does not create any issue since the tests
successfully uncompress that buffer to the original data.

The longer the compressed waveform is, the more likely some compression
is happening, thus avoiding the bug that happens only when absolutely no
compression is achieved.

The fix makes it impossible for this to happen again, at the price of a
slight slow down. A different fix may have no slow down, leaving the
code more brittle than it is now.