Tentative fix to RangeForWrapperBox (util::range_for) per issue #15400 .
I was saving the begin and end iterators from the (potentially temporary) range into my wrapper class.
Chances are that those iterators still need the range to be around to work (as they have the right to demand).
Sometimes the temporary was disappearing too early, and the iterators would go berserk.